+++ src/dashboardWebView/index.tsx
@@ -31,7 +31,7 @@ declare global {
       getCardFooter: (filePath: string, data: any) => Promise<string | undefined>;
       // 8.5.0 extension points
       getCardTitle: (filePath: string, data: any) => Promise<string | undefined>;
-      getcardDescription: (filePath: string, data: any) => Promise<string | undefined>;
+      getCardDescription: (filePath: string, data: any) => Promise<string | undefined>;
       getCardTags: (filePath: string, data: any) => Promise<string | undefined>;
       getCardDate: (filePath: string, data: any) => Promise<string | undefined>;
       getCardStatus: (filePath: string, data: any) => Promise<string | undefined>;