+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [#566](https://github.com/estruyf/vscode-front-matter/issues/566): Keep the panel context on the live preview
 - [#568](https://github.com/estruyf/vscode-front-matter/issues/568): Update the preview URL if the slug changes
 - [#586](https://github.com/estruyf/vscode-front-matter/issues/586): Allow to specify the content card fields
+- [#588](https://github.com/estruyf/vscode-front-matter/issues/588): Added extensibility support to override card fields
 
 ### ⚡️ Optimizations